auth_service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
itsdangerous>=2.1.2
aiofiles>=23.2.0

